Refactor task into it's own module and create ToDoList class
